Message section done

+++ base/views.py
@@ -74,6 +74,7 @@ def home(request):
 def room(request, pk):
     room = Room.objects.get(id=pk)
     room_messages = room.message_set.all().order_by('-created')
+    participants = room.participants.all()
 
     if request.method == 'POST':
         message = Message.objects.create(
@@ -81,9 +82,10 @@ def room(request, pk):
             room = room,
             body = request.POST.get('body')
         )
+        room.participants.add(request.user)
         return redirect('room', pk=room.id)
 
-    context = {'room' : room, 'room_messages' : room_messages}
+    context = {'room' : room, 'room_messages' : room_messages, 'participants' : participants}
     return render(request, 'base/room.html', context)
 
 @login_required(login_url="login")
@@ -116,7 +118,20 @@ def updateRoom(request, pk):
 @login_required(login_url="login")
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not Allowed!!!')
     if request.method == "POST":
         room.delete()
         return redirect('home')
     return render(request, 'base/delete.html', {'obj' : room})
+
+@login_required(login_url="login")
+def deleteMessage(request, pk):
+    message = Message.objects.get(id=pk)
+    if request.user != message.user:
+        return HttpResponse('You are not Allowed!!!')
+    if request.method == "POST":
+        message.delete()
+        return redirect('home')
+    return render(request, 'base/delete.html', {'obj' : message})
